+++ QueryRenderer/QueryRenderManager.h
@@ -203,7 +203,7 @@ class QueryRenderManager {
   PngData renderToPng(int compressionLevel = -1);
 
   // get the id at a specific pixel
-  unsigned int getIdAt(int x, int y);
+  int64_t getIdAt(int x, int y);
 
   // CROOT - the following is a debug function. Remove when ready for deployment
   GLFWwindow* getWindow() {
